snapshot of project "cdk", label t20110517

+++ cli/cdkdialog.c
@@ -1,4 +1,4 @@
-/* $Id: cdkdialog.c,v 1.10 2008/11/04 23:36:39 tom Exp $ */
+/* $Id: cdkdialog.c,v 1.11 2011/05/16 22:47:37 tom Exp $ */
 
 #include <cdk_test.h>
 
@@ -91,7 +91,7 @@ int main (int argc, char **argv)
    {
       /* Split the message up. */
       messageList = CDKsplitString (message, '\n');
-      messageLines = CDKcountStrings (messageList);
+      messageLines = (int) CDKcountStrings (messageList);
    }
 
    /* Set up the buttons for the dialog box. */
@@ -105,7 +105,7 @@ int main (int argc, char **argv)
    {
       /* Split the button list up. */
       buttonList = CDKsplitString (buttons, '\n');
-      buttonCount = CDKcountStrings (buttonList);
+      buttonCount = (int) CDKcountStrings (buttonList);
    }
 
    /* Set up CDK. */
