+++ cli/cdkitemlist.c
@@ -1,4 +1,4 @@
-/* $Id: cdkitemlist.c,v 1.10 2008/11/04 23:36:39 tom Exp $ */
+/* $Id: cdkitemlist.c,v 1.11 2011/05/16 22:53:13 tom Exp $ */
 
 #include <cdk_test.h>
 
@@ -104,7 +104,7 @@ int main (int argc, char **argv)
    {
       /* Split the itemlist lines up. */
       itemlistList = CDKsplitString (list, '\n');
-      itemlistLines = CDKcountStrings (itemlistList);
+      itemlistLines = (int) CDKcountStrings (itemlistList);
    }
 
    /* Start curses. */
@@ -155,7 +155,7 @@ int main (int argc, char **argv)
    if (buttons != 0)
    {
       buttonList = CDKsplitString (buttons, '\n');
-      buttonCount = CDKcountStrings (buttonList);
+      buttonCount = (int) CDKcountStrings (buttonList);
 
       /* We need to create a buttonbox widget. */
       buttonWidget = newCDKButtonbox (cdkScreen,
